refactor: Remove numeric and braces
Numeric values do not need to be treated differently than string ones and especially since it's not needed to use quotes it makes it problematic to check what is a number.

I also removed any kind of braces, parens etc. since I would need to fix the tests and they no longer make sense with this version of directives.